node-finder fdlimit setting fix
